+++ src/distro/index.ts
@@ -9,4 +9,8 @@ export type {
 } from "./types.js";
 export { MICROSOFT_OPENTELEMETRY_VERSION } from "./types.js";
 
-export { useMicrosoftOpenTelemetry, shutdownMicrosoftOpenTelemetry } from "./distro.js";
+export {
+  useMicrosoftOpenTelemetry,
+  shutdownMicrosoftOpenTelemetry,
+  whenGenAIInstrumentationsReady,
+} from "./distro.js";